grpc-service/requirements.txt
fastapi
uvicorn
aiokafka
redis
grpcio
grpcio-tools
# sqlite3
python-jose
protobuf==5.27.2